NetService removal refactoring
Moved GetNodeSegmentIds out of NetService into ExtNodeManager.
Moved GetInitialSegment out of NetService into ExtNodeManager.
Moved ClockDirection out of INetService.cs into its own file in TrafficManager.Impl Namespace.
