+++ TLM/Benchmarks/GetNodeSegmentIdsEnumerablePerfTests.cs
@@ -1,6 +1,6 @@
 using BenchmarkDotNet.Attributes;
-using GenericGameBridge.Service;
 using System.Collections.Generic;
+using TrafficManager.Manager.Impl;
 
 namespace Benchmarks {
     public class GetNodeSegmentIdsEnumerablePerfTests {
