+++ TLM/TLM/Manager/Impl/ClockDirection.cs
@@ -1,0 +1,6 @@
+namespace TrafficManager.Manager.Impl {
+    public enum ClockDirection {
+        Clockwise,
+        CounterClockwise,
+    }
+}
