+++ TLM/TLM/Manager/Impl/GetNodeSegmentIdsEnumerator.cs
@@ -3,7 +3,7 @@ using System;
 using System.Collections;
 using System.Collections.Generic;
 
-namespace GenericGameBridge.Service {
+namespace TrafficManager.Manager.Impl {
     /// <summary>
     /// Mutable struct enumerator to be returned from the GetNodeSegmentIds method.
     /// This implementation is just for perf optimizations and should be handled with care since it is a mutable struct!
