+++ TLM/TMPE.GenericGameBridge/TMPE.GenericGameBridge.csproj
@@ -76,9 +76,7 @@
     </Compile>
     <Compile Include="Factory\IServiceFactory.cs" />
     <Compile Include="Service\GetSegmentLaneIdsEnumerable.cs" />
-    <Compile Include="Service\GetNodeSegmentIdsEnumerable.cs" />
     <Compile Include="Service\GetSegmentLaneIdsEnumerator.cs" />
-    <Compile Include="Service\GetNodeSegmentIdsEnumerator.cs" />
     <Compile Include="Service\INetService.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="Service\LaneIdAndIndex.cs" />
